requirements.txt
Django>=4.2
djangorestframework
python-dotenv
drf-spectacular
drf-standardized-errors
djangorestframework-simplejwt
django-filter>=21.1
pytest
pytest-django
pytest-factoryboy
faker